1) fixed the FXSR detection in the FPU code. 2) Re-organized the par_lib files. 3) Moved the OpenMP applications to implementation/tests/
