+++ src/components/implementation/no_interface/acap_ping/Makefile
@@ -2,7 +2,7 @@ C_OBJS=ping.o
 ASM_OBJS=
 COMPONENT=acap_pingp.o
 INTERFACES=
-DEPENDENCIES=acap_pong sched printc par_mgr mem_mgr_large valloc par_mgr_intra
+DEPENDENCIES=par_mgr acap_pong sched printc par_mgr mem_mgr_large valloc 
 IF_LIB=
 
 include ../../Makefile.subsubdir
