--- src/components/implementation/omp/Makefile
@@ -1,4 +1,0 @@
-INTERFACES=omp
-DEPENDENCIES=mem_mgr_large valloc printc sched lock par_mgr_intra
-
-include ../Makefile.subdir
