--- src/components/implementation/omp/default/Makefile
@@ -1,11 +1,0 @@
-C_OBJS=openmp.o
-ASM_OBJS=
-COMPONENT=omp_comp.o
-INTERFACES=omp
-DEPENDENCIES=mem_mgr_large sched valloc printc par_mgr_intra timed_blk lock
-IF_LIB=../complib.o
-
-include ../../Makefile.subsubdir
-
-CFLAGS += -fopenmp
-
