+++ src/components/implementation/par_mgr/default/Makefile
@@ -1,7 +1,7 @@
 C_OBJS=par_mgr.o
 ASM_OBJS=
 COMPONENT=parmgr.o
-INTERFACES=par_mgr par_mgr_intra
+INTERFACES=par_mgr
 DEPENDENCIES=sched printc valloc mem_mgr_large
 IF_LIB=
