+++ src/components/implementation/tests/omp_fft/Makefile
@@ -1,9 +1,9 @@
 C_OBJS=omp_fft.o
 ASM_OBJS=
 COMPONENT=fft.o
-INTERFACES=omp
-DEPENDENCIES=mem_mgr_large sched valloc printc par_mgr_intra timed_blk lock
-IF_LIB=../complib.o /usr/lib/libm.a
+INTERFACES=
+DEPENDENCIES=par_mgr mem_mgr_large sched valloc printc timed_blk lock
+IF_LIB=/usr/lib/libm.a
 
 include ../../Makefile.subsubdir
