+++ src/components/implementation/tests/omp_jacobi/Makefile
@@ -1,9 +1,9 @@
 C_OBJS=omp_jacobi.o
 ASM_OBJS=
 COMPONENT=cos_jacobi.o
-INTERFACES=omp
-DEPENDENCIES=mem_mgr_large sched valloc printc par_mgr_intra timed_blk lock
-IF_LIB=../complib.o /usr/lib/libm.a
+INTERFACES=
+DEPENDENCIES=par_mgr mem_mgr_large sched valloc printc timed_blk lock
+IF_LIB=/usr/lib/libm.a
 
 include ../../Makefile.subsubdir
