+++ src/components/implementation/tests/omp_lu/Makefile
@@ -1,9 +1,8 @@
 C_OBJS=omp_lu.o
 ASM_OBJS=
 COMPONENT=cos_lu.o
-INTERFACES=omp
-DEPENDENCIES=mem_mgr_large sched valloc printc par_mgr_intra timed_blk lock
-IF_LIB=../complib.o
+INTERFACES=
+DEPENDENCIES=par_mgr mem_mgr_large sched valloc printc timed_blk lock
 
 include ../../Makefile.subsubdir
