+++ src/components/implementation/tests/unit_par/Makefile
@@ -1,0 +1,10 @@
+C_OBJS=openmp.o
+ASM_OBJS=
+COMPONENT=omp_comp.o
+INTERFACES=
+DEPENDENCIES=par_mgr mem_mgr_large sched valloc printc lock timed_blk
+
+include ../../Makefile.subsubdir
+
+CFLAGS += -fopenmp
+
