--- src/components/interface/omp/Makefile
@@ -1,5 +1,0 @@
-LIB_OBJS=
-LIBS=$(LIB_OBJS:%.o=%.a)
-ASM_STUBS=
-
-include ../Makefile.subdir
