--- src/components/interface/omp/omp_cos.h
@@ -1,6 +1,0 @@
-#ifndef OMP_COS_H
-#define OMP_COS_H
-
-/* Nothing here */
-
-#endif /* OMP_COS_H */
