--- src/components/interface/omp/stubs/s_stub.S
@@ -1,10 +1,0 @@
-/**
- * Copyright 2011 by Gabriel Parmer, gparmer@gwu.edu
- *
- * Redistribution of this file is permitted under the GNU General
- * Public License v2.
- */
-
-#include <cos_asm_server_stub.h>
-.text	
-
