+++ src/components/interface/par_mgr/Makefile
@@ -1,6 +1,6 @@
-LIB_OBJS=ainv_lib.o
+LIB_OBJS=parlib.o
 LIBS=$(LIB_OBJS:%.o=%.a)
-ASM_STUBS=s_stubam.o
-DEPENDENCIES=sched
+ASM_STUBS=s_stub.o
+DEPENDENCIES=mem_mgr_large sched lock
 
 include ../Makefile.subdir
