+++ src/components/interface/par_mgr/stubs/s_stub.S
@@ -15,3 +15,13 @@ cos_asm_server_stub_spdid(acap_cli_lookup)
 cos_asm_server_stub_spdid(acap_cli_lookup_ring)
 cos_asm_server_stub_spdid(acap_srv_lookup)
 cos_asm_server_stub_spdid(acap_srv_lookup_ring)
+
+cos_asm_server_stub_spdid(par_create)
+cos_asm_server_stub_spdid(par_acap_lookup)
+cos_asm_server_stub_spdid(par_ring_lookup)
+cos_asm_server_stub_spdid(par_acap_get_barrier)
+cos_asm_server_stub_spdid(par_srv_thd_num_lookup)
+cos_asm_server_stub_spdid(par_srv_acap_lookup)
+cos_asm_server_stub_spdid(par_srv_ring_lookup)
+cos_asm_server_stub_spdid(par_parent_lookup)
+
