--- src/components/interface/par_mgr_intra/Makefile
@@ -1,6 +1,0 @@
-LIB_OBJS=par_lib.o
-LIBS=$(LIB_OBJS:%.o=%.a)
-ASM_STUBS=s_stubam.o
-DEPENDENCIES=mem_mgr_large valloc lock
-
-include ../Makefile.subdir
