--- src/components/interface/par_mgr_intra/par_lib.c
@@ -1,6 +1,0 @@
-#include <cos_component.h>
-#include <print.h>
-#include <cos_alloc.h>
-#include <par_mgr_intra.h>
-
-struct par_thd_info *__par_thd_info[MAX_NUM_THREADS]; // TODO: replace with cvect
