--- src/components/interface/par_mgr_intra/stubs/s_stub.S
@@ -1,20 +1,0 @@
-/**
- * Copyright 2013 by Qi Wang, interwq@gwu.edu
- *
- * Redistribution of this file is permitted under the GNU General
- * Public License v2.
- */
-
-#include <cos_asm_server_stub.h>
-
-.text
-
-cos_asm_server_stub_spdid(par_create)
-cos_asm_server_stub_spdid(par_acap_lookup)
-cos_asm_server_stub_spdid(par_ring_lookup)
-cos_asm_server_stub_spdid(par_acap_get_barrier)
-cos_asm_server_stub_spdid(par_srv_thd_num_lookup)
-cos_asm_server_stub_spdid(par_srv_acap_lookup)
-cos_asm_server_stub_spdid(par_srv_ring_lookup)
-cos_asm_server_stub_spdid(par_parent_lookup)
-
