+++ src/kernel/include/fpu.h
@@ -6,7 +6,7 @@
 #define ENABLE            1
 #define DISABLE           0
 #define FPU_DISABLED_MASK 0x8
-#define FXSR              1<<24
+#define FXSR              (1<<24)
 
 PERCPU_DECL(int, fpu_disabled);
 PERCPU_EXTERN(fpu_disabled);
@@ -34,19 +34,23 @@ static inline void fxsave(struct thread*);
 static inline void fxrstor(struct thread*);
 static inline unsigned long fpu_read_cr0(void);
 static inline void fpu_set(int);
-static inline u64_t fpu_get_cpuid(void);
+static inline int fpu_get_cpuid(void);
 static inline int fpu_check_fxsr(void);
 
 #ifdef FPU_ENABLED
-static inline u64_t
+static inline int
 fpu_get_cpuid(void)
 {
-        u64_t cpuid;
+        int cpuid;
 
         asm volatile("mov $1, %%eax\n\t"
                      "cpuid\n\t"
                      "movl %%edx, %0"
-                     : "=m" (cpuid) : : "eax", "edx");
+                     : "=m" (cpuid) 
+		     : 
+		     : "eax", "ebx", "ecx", "edx");
+
+	/* printk("cpu %d cpuid_edx %x\n", get_cpuid(), cpuid); */
 
         return cpuid;
 }
@@ -54,11 +58,12 @@ fpu_get_cpuid(void)
 static inline int
 fpu_check_fxsr(void)
 {
-        u64_t cpuid;
+        int cpuid;
         int fxsr_status;
 
         cpuid = fpu_get_cpuid();
-        fxsr_status = cpuid & FXSR; /* SSE presents if bit 25 is set, more information refer to CPUID instruction */
+	/* fxsr is the 25th bit (start from bit 1) in EDX. So FXSR is 1<<24. */
+        fxsr_status = ((cpuid & FXSR) != 0) ? 1 : 0 ;
 
         return fxsr_status;
 }
@@ -68,7 +73,7 @@ fpu_init(void)
 {
         *PERCPU_GET(fpu_support_fxsr) = fpu_check_fxsr();
 
-        if (!*PERCPU_GET(fpu_support_fxsr)) printk("fpu doesn't support fxsave/fxrstor, use fsave/frstr instead\n");
+        if (!*PERCPU_GET(fpu_support_fxsr)) printk("Core %d: fpu doesn't support fxsave/fxrstor, use fsave/frstr instead\n", get_cpuid());
 
         fpu_set(DISABLE);
 	*PERCPU_GET(fpu_disabled) = 1;
@@ -240,7 +245,7 @@ static inline void fxsave(struct thread *thd) { return; }
 static inline void fxrstor(struct thread *thd) { return; }
 static inline unsigned long fpu_read_cr0(void) { return 0; };
 static inline void fpu_set(int status) { return; }
-static inline u64_t fpu_get_cpuid(void) { return 0; }
+static inline int fpu_get_cpuid(void) { return 0; }
 static inline int fpu_check_fxsr(void) { return 0; }
 #endif
